moved add Friend button to component
