+++ ChainLink-Client/src/components/ProfileModal.tsx
@@ -1,9 +1,9 @@
 import { polygon } from 'leaflet';
-import React, { useState, useContext, useEffect } from 'react';
+import React from 'react';
 import '../styles/components/profile-modal.css';
 import { Tooltip } from 'react-tooltip';
-import { gql, useQuery, useMutation } from '@apollo/client';
-import { AuthContext } from '../context/auth';
+import { gql, useQuery } from '@apollo/client';
+import AddFriendButton from './AddFriendButton'; // Import the new AddFriendButton
 
 interface ProfileModalProps {
     user: any | null;
@@ -30,29 +30,7 @@ const FETCH_USER_QUERY = gql`
   }
 `;
 
-// Check if a friend request is already pending
-const CHECK_FRIEND_STATUS_QUERY = gql`
-  query checkFriendStatus($senderId: ID!, $recipientId: ID!) {
-    checkFriendStatus(senderId: $senderId, recipientId: $recipientId) {
-      status
-    }
-  }
-`;
-
-// Mutation to send a friend request
-const ADD_FRIEND_MUTATION = gql`
-  mutation addFriend($senderId: ID!, $recipientId: ID!) {
-    addFriend(senderId: $senderId, recipientId: $recipientId) {
-      id
-      status
-      created_at
-    }
-  }
-`;
-
 export const ProfileModal: React.FC<ProfileModalProps> = ({ user }) => {
-    const { user: currentUser } = useContext(AuthContext);
-    const [friendStatus, setFriendStatus] = useState<'add' | 'pending'>('add');
     const foreColor = window.getComputedStyle(document.documentElement).getPropertyValue('--primary-color');
 
     const { loading: userLoading, error, data: userData } = useQuery(FETCH_USER_QUERY, {
@@ -61,45 +39,7 @@ export const ProfileModal: React.FC<ProfileModalProps> = ({ user }) => {
         },
     });
 
-    // Query to check the friend request status
-    const { loading: friendStatusLoading, data: friendStatusData } = useQuery(CHECK_FRIEND_STATUS_QUERY, {
-        variables: {
-            senderId: currentUser?.id,
-            recipientId: userData?.getUser?.id,
-        },
-        skip: !currentUser || !userData?.getUser?.id, // Skip if IDs are not available
-        onCompleted: (data) => {
-            if (data.checkFriendStatus && data.checkFriendStatus.status === 'pending') {
-                setFriendStatus('pending');
-            }
-        }
-    });
-
-    const [addFriend] = useMutation(ADD_FRIEND_MUTATION, {
-        onCompleted: () => {
-            setFriendStatus('pending');
-        },
-        onError: (err) => {
-            console.error("Error adding friend:", err);
-        }
-    });
-
-    const handleAddFriendClick = () => {
-        if (!currentUser || !currentUser.id || !userData || !userData.getUser.id) {
-            console.error("User IDs not available for friend request.");
-            return;
-        }
-
-        // Send a friend request to the server
-        addFriend({
-            variables: {
-                senderId: currentUser.id,
-                recipientId: userData.getUser.id,
-            }
-        });
-    };
-
-    if (userLoading || friendStatusLoading) {
+    if (userLoading) {
         return (
             <Tooltip 
                 anchorSelect={"#profile-modal-anchor-" + user} 
@@ -151,13 +91,7 @@ export const ProfileModal: React.FC<ProfileModalProps> = ({ user }) => {
                     <span className="profile-modal-descriptor-right">{userData.getUser.eventsHosted.length + " Rides Joined"}</span>
                 </div>
                 <div className="profile-modal-actions">
-                    <button
-                        onClick={handleAddFriendClick}
-                        className="add-friend-button"
-                        disabled={friendStatus === 'pending'}
-                    >
-                        {friendStatus === 'pending' ? 'Pending' : 'Add Friend'}
-                    </button>
+                    <AddFriendButton recipientId={userData.getUser.id} />
                 </div>
             </div>
         </Tooltip>
